fixed logging

+++ ArduPlane/mode_guided.cpp
@@ -28,7 +28,7 @@ bool ModeGuided::_enter()
     currentBearing = wrap_360_cd(plane.ahrs.yaw_sensor) + targetAngle * 100;
     minAlt = constrain_float(relative_altitude_cm - plane.g.hm_alt_diff, plane.g.hm_min_alt, relative_altitude_cm);
 
-    hal.console->printf("Hd: %ld ; Tgt_Hd %ld\n", wrap_360_cd(plane.ahrs.yaw_sensor), currentBearing);
+    hal.console->printf("Hd: %ld ; Tgt_Hd %ld\n", (long int)wrap_360_cd(plane.ahrs.yaw_sensor), (long int)currentBearing);
     hal.console->printf("min altitude: %f \n", minAlt);
 
     stopRoll = false;
